fix(api): move delete org
Ref. metriport/metriport-internal#187
